Indian-Market_Algobot/requirements.txt
# Requirements for algo-trading project

# Core dependencies
pandas
numpy
matplotlib

# Broker API and WebSocket
smartapi-python==1.4.8
websocket-client==1.8.0
pyotp==2.8.0

# HTTP and Data Processing
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Environment Variables
python-dotenv==1.0.0

# Logging
logzero==1.7.0

# API Server (Dashboard)
fastapi
uvicorn[standard]

# Cryptography and Security
cryptography==43.0.1
certifi==2023.5.7

# WebSocket Dependencies
autobahn==24.4.2
txaio==23.1.1

# Required by Dependencies
attrs==23.1.0
Automat==22.10.0
cffi>=1.17.0
charset-normalizer==3.1.0
constantly==15.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pycparser==2.21
six==1.16.0
typing_extensions==4.6.3
urllib3==2.0.3
zope.interface>=7.1.0
